Move UserCard to User Page
